refactor(examples): Inline shared PPO scaffolding
Drop the `#[path = "common/ppo_cartpole.rs"]` include and inline the
PPO-on-CartPole scaffolding into each consumer example so each is a
self-contained runnable file. The two viz tiers differed only in their
env wrappers (TuiEnvTap vs. RecordingTap); the agent, value network,
and hyperparams are now duplicated verbatim, trading a small amount of
duplication for examples that compile without the path-attribute trick.

CRITICAL: The turbofish
`train_discrete::<Be, _, _, _, _, CartPoleAction, _, 1, 1, 2>` and the
`PpoValue<B, 2>` bound on `ValueMlp<B: AutodiffBackend>` carry
type-level facts (scalar reward, rank-1 state/action tensors, 2
actions) that must stay byte-identical at every inlined call site or
the agent will fail to build.
